1. The configuration way has been changed. Now constructor receives a configuration object instead of just `entitySuffix` param.
First of all, it is a breaking change. So, it is a required step and it will be better to do the change now until the library is not used in a lot of production environments yet.

It has been done because we will have more configuration params in the future and we don't want to have a heap of optional function arguments. In accordance with 'The Open Close Principle' of S.O.L.I.D., we decided to use an object for configuring the naming strategy.

2. In-Db suffixes implemented.
In-Db suffix is a suffix that should be added as is to the table name and should not be pluralized. But the name before the suffix should be pluralized.

3. Existing .entitySuffix param and new .inDbSuffix got and ability be receive multiple values. So now you can have multiple suffixed of each kind.