1. Building names multi-column indexes and constraints changed. Columns sorting removed.
In the previous version, columns were sorted. The reason why it was done was thinging: 'TypeORM can change columns order under the hood and we don't need to change the index names due to this.".

But this assumption was wrong. The columns order is very important when we are creating an index. I realized it only now when faced with performance issues related to columns order. So now I understand that TypeORM should not change the index order.

Actually the index name doesn't affect anything, however different columns order in the index and in the name of the index can confuse.